Mauro:  Extranet de Asociados AMAT (int-027) Revisión 0.2

+++ branches/int_027_v2/apps/frontend/modules/archivos_d_g/templates/indexSuccess.php
@@ -47,7 +47,7 @@ endif; ?>
 				<div style="float:left;" class="paginado"><?php echo test_pager($pager, $orderBy, $sortType, $redireccionGrupo) ?></div>
 			<?php endif; ?>
 			<span class="info" style="float: left;">Hay <?php echo $cantidadRegistros ?> Registro/s <?php if ($cajaBsq) echo " con la palabra '".$cajaBsq."'" ?> </span> 
-			<?php if(validate_action('alta')):?>
+			<?php if(validate_action('alta') && $carga == 1):?>
 			<input type="button" onclick="javascript:location.href='<?php echo url_for('archivos_d_g/nueva?'.$redireccionGrupo) ?>';" style="float: right;" value="Nuevo Archivo" name="newNews" class="boton"/>
 			<?php endif;?>
 		</div>
@@ -102,14 +102,14 @@ endif; ?>
 						<?php endif; ?>
 					</td>
 					<td valign="center" align="center">
-					<?php if(validate_action('modificar') || $valor->getOwnerId() == $sf_user->getAttribute('userId')):?>
+					<?php if(validate_action('modificar') || $valor->getOwnerId() == $sf_user->getAttribute('userId') && $carga == 1):?>
 						<a href="<?php echo url_for('archivos_d_g/editar?id=' . $valor->getId().'&'.$redireccionGrupo) ?>">
 							<?php echo image_tag('show.png', array('height' => 20, 'width' => 17, 'border' => 0, 'title' => 'Ver')) ?>
 						</a>
 						<?php endif;?>
 					</td>
                                           <td valign="center" align="center">
-                                          <?php if(validate_action('baja')  || $valor->getOwnerId() == $sf_user->getAttribute('userId')):?>
+                                          <?php if(validate_action('baja')  || $valor->getOwnerId() == $sf_user->getAttribute('userId') && $carga == 1):?>
                                                 <?php echo link_to(image_tag('borrar.png', array('title'=>'Borrar','alt'=>'Borrar','width'=>'20','height'=>'20','border'=>'0')), 'archivos_d_g/delete?id='.$valor->getId().'&'.$redireccionGrupo, array('method'=>'delete','confirm'=>'Confirma la eliminaci&oacute;n del registro?')) ?>
                                           <?php endif; ?>
                                           </td>
@@ -139,7 +139,7 @@ endif; ?>
 			<?php endif; ?>
 			
 			<span class="info" style="float: left;">Hay <?php echo $cantidadRegistros ?> Registro/s</span>
-            <?php if(validate_action('alta')):?>
+                         <?php if(validate_action('alta') && $carga == 1):?>
 			<input type="button" onclick="javascript:location.href='<?php echo url_for('archivos_d_g/nueva?'.$redireccionGrupo) ?>';" style="float: right;" value="Nuevo Archivo" name="newNews" class="boton"/>
 			<?php endif;?>
